+++ src/main/java/lagatrix/server/connection/requester/RequestCPU.java
@@ -1,6 +1,7 @@
 package lagatrix.server.connection.requester;
 
 import lagatrix.server.connection.communicators.AESCommunicator;
+import lagatrix.server.entities.actions.ActionsEnum;
 import lagatrix.server.entities.connection.Request;
 import lagatrix.server.entities.connection.Response;
 import lagatrix.server.exceptions.LagatrixException;
@@ -16,30 +17,36 @@ import lagatrix.server.tools.command.CommandExecutor;
 public class RequestCPU extends RequestManager {
 
     private CPUManager manager;
-    
-    public RequestCPU(AESCommunicator communicator, CommandExecutor executor) {
+
+    public RequestCPU(AESCommunicator communicator, CommandExecutor executor, boolean isRaspberry) {
         super(communicator, executor);
+        manager = CPUManager(executor, isRaspberry);
     }
-    
+
     @Override
     public void determineRequest(Request request) throws LagatrixException {
-        Response response;
+        Response response = new Response();
         char c;
-        
+
         // Determine if obtain CPU info, temperature or use.
-        if (request.getParams().length > 0){
-            c = (char) request.getParams()[0];
-            
-            if (c == 'U' || c == 'u') {
-                response = new Response((Float) manager.obtainUse());
+        if (request.getAction() == ActionsEnum.GET) {
+            if (request.getParams().length > 0) {
+                c = (char) request.getParams()[0];
+
+                if (c == 'U' || c == 'u') {
+                    response.setResponse((Float) manager.obtainUse());
+                } else if (c == 'T' || c == 't') {
+                    response.setResponse((Float) manager.obtainTemperature());
+                }
             } else {
-                response = new Response((Float) manager.obtainTemperature());
+                response.setResponse(manager.obtainCPU());
             }
-        } else {
-            response = new Response(manager.obtainCPU());
         }
-        
-        response.setCorrectResult(true);
+
         communicator.sendResponse(response);
+    }
+
+    private CPUManager CPUManager(CommandExecutor executor, boolean raspberry) {
+        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
     }
 }