+++ src/main/java/lagatrix/server/entities/connection/Response.java
@@ -17,17 +17,14 @@ public class Response implements Serializable {
         this(null, false);
     }
 
-    public Response(Object response) {
-        this(response, true);
-    }
-
-    public Response(boolean correctResult) {
-        this(null, correctResult);
-    }
-
     public Response(Object response, boolean correctResult) {
         this.correctResult = correctResult;
         this.response = response;
+    }
+
+    public void setResponse(Object response) {
+        this.response = response;
+        setCorrectResult(true);
     }
 
     public void setCorrectResult(boolean correctResult) {